requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
flask-cors
flask-limiter
python-dateutil
python-dotenv
psycopg2-binary
gunicorn
apscheduler
